Added UI functionality
Added CompletionUI.cs for controlling whether or not a dialogue box appears when the user completes the circuit. Right now, I just have a one-line addition to Battery.cs that makes it pop up when the battery is connected, but this functionality can be moved to CircuitLab.cs once we actually have checks for a completed circuit.